refact(response): move response, encoder and error into response package
