+++ response/error_test.go
@@ -1,7 +1,7 @@
 // Copyright 2017 Axel Etcheverry. All rights reserved.
 // Use of this source code is governed by a MIT
 // license that can be found in the LICENSE file.
-package server
+package response
 
 import (
 	"testing"
