+++ response/register_test.go
@@ -2,11 +2,12 @@
 // Use of this source code is governed by a MIT
 // license that can be found in the LICENSE file.
 
-package encoder
+package response
 
 import (
 	"testing"
 
+	"github.com/euskadi31/go-server/response/encoder"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -14,7 +15,7 @@ func TestRegister(t *testing.T) {
 
 	assert.Equal(t, 0, len(encoders))
 
-	provider := &MockEncoder{}
+	provider := &encoder.MockEncoder{}
 
 	provider.On("MimeType").Return("application/json")
 
@@ -22,5 +23,5 @@ func TestRegister(t *testing.T) {
 
 	assert.Equal(t, 1, len(encoders))
 
-	encoders = make(map[string]Encoder)
+	encoders = make(map[string]encoder.Encoder)
 }
